Make Parcel an interface
Currently, there is a class named Parcel and the field Action determines
the type of the Parcel.
This patch makes Parcel an interface(actually it's an abstract class),
and it changes Actions to Parcels. So the terms related Action are all
removed(the only remain action is a field name).

And it extracts fromJSON of Parcel and SignedParcel to avoid cyclic
referencing.

And it also makes SignedParcel use the implementation of the unsigned
Parcel to create an encode object and json object.