Skip Memory pool's memory limit test
Foundry has a module system to make a user be able to create its logic
using it. If the user needs a custom asset, they will implement it
using the module. So Foundry does not need the Asset system.

Memory pool's memory limit test was using the AssetMint transaction to
create a transaction whose size is vast. After removing the AssetMint
transaction, we have only small-sized transactions. We need to send
more than 10,000 transactions to over the limit. Since sending 10,000
transactions takes too much time, let's skip this test and enable it later.