Calculate the transactions root with the transactions on the block
Currently, the transactions root was the skewed merkle root of all
accumulated transactions from the genesis block. It changes the
definition of the transactions root to the skewed merkle root of
transactions in the block.

This change makes the node be able to verify a block's transactions
root without the parent header.