Reduce the number of db transactions of PeerDb
Currently, the address is saved in the db immediately. It increases the
disk usage on the booting since there will many network update.

This patch changes PeerDb to update disk per 10 times. It also makes the
db store all peers in the same field. The constant "peer-list" is used
for the key. And dedicated column for the peer list is removed. The
addresses will be stored in the extra column.