Implement Coordinator interface
This commit introduces new interface called filter_transactions, which
should accept all transactions in a mempool as their inputs.
Note that it should accept transactions with metadata, such as their
origin.
It should filter away old transactions and return transactions that
should remain in the mempool.

The role of check_transaction is a guard of mempool, to prevent spam
transactions.
check_transaction should support options more than just recheck for
flexibility. For example, we can skip more checks when we are checking
local transactions.

open_block, execute_transactions, and close_block are to support
streaming feature.
Also, we decided to use context as an argument of the calls, instead of
keeping context variable inside Coordinator.

The miner's block preparation logic should be changed accordingly.