Remove CodeChainMachine
CodeChainMachine was designed to abstract the rule of transaction and
state change, but it will be moved to the module. So this abstraction in
the core module is not necessary anymore.

There are five methods in CodeChainMachine.
- balance, add_balance, increase_term_id are not used.
- verify_transaction does nothing.
The only meaningful method is verify_transaction_seal. This commit makes
the caller call the internal code directly.