Move simple.rs from /tests to /tests/simple
When building mold branch on foundry, we met the warning "file foundry
to be present in multiple build targets". It seems to occur if there is
another target in the directory where the target is located, when
specifying a build target.

I resolved the warning by creating a new directory `tests/simple` and
moving `tests/simple.rs` to `tests/simple/simple.rs`.