Change fetch_transactions_for_block into prepare_transactions_with_result
If we use TxFIlter::fetch_transactions_for_block, then validators have
to execute the same transactions twice.
 This is because preparing transactions inherently requires
execution of transactions.
We should change the interface in order to avoid the double execution.
The miner's block preparation logic is changed accordingly.
TransactionWithGas type is removed as well since the host doesn't need the
concept of gas anymore.